requirements-dev.txt
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0.5
django-filter==2.0.0
djangorestframework==3.9.0
Pillow==5.3.0
python-decouple==3.1
pytz==2018.7
static3==0.7.0
style==1.1.0
update==0.0.1
